Extract sidebar user menu into a reusable component
