+++ apps/web/src/components/layout/sidebar-user-menu.tsx
@@ -1,0 +1,81 @@
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuGroup,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@vita-os/ui/components/dropdown-menu";
+import {
+  SidebarMenu,
+  SidebarMenuButton,
+  SidebarMenuItem,
+} from "@vita-os/ui/components/sidebar";
+import { ChevronsUpDown, LogOut } from "lucide-react";
+
+interface SidebarUser {
+  name?: string | null;
+  email?: string | null;
+}
+
+interface SidebarUserMenuProps {
+  user?: SidebarUser | null;
+  onSignOut: () => void;
+}
+
+export function SidebarUserMenu({ user, onSignOut }: SidebarUserMenuProps) {
+  const accountName = user?.name ?? user?.email ?? "Account";
+  const initial = (user?.name?.[0] ?? user?.email?.[0] ?? "?").toUpperCase();
+
+  return (
+    <SidebarMenu>
+      <SidebarMenuItem>
+        <DropdownMenu>
+          <DropdownMenuTrigger
+            render={<SidebarMenuButton size="lg" tooltip={accountName} />}
+          >
+            <div className="flex size-8 items-center justify-center rounded-lg bg-sidebar-primary text-sidebar-primary-foreground">
+              {initial}
+            </div>
+            <div className="grid flex-1 text-left text-sm leading-tight">
+              <span className="truncate font-medium">{accountName}</span>
+              {user?.name && (
+                <span className="truncate text-xs text-muted-foreground">
+                  {user.email}
+                </span>
+              )}
+            </div>
+            <ChevronsUpDown className="ml-auto size-4" />
+          </DropdownMenuTrigger>
+          <DropdownMenuContent
+            className="w-[--radix-dropdown-menu-trigger-width] min-w-56"
+            side="top"
+            align="start"
+            sideOffset={4}
+          >
+            <DropdownMenuGroup>
+              <DropdownMenuLabel className="font-normal">
+                <div className="flex flex-col gap-1">
+                  {user?.name && (
+                    <p className="text-sm font-medium leading-none">
+                      {user.name}
+                    </p>
+                  )}
+                  <p className="text-xs leading-none text-muted-foreground">
+                    {user?.email}
+                  </p>
+                </div>
+              </DropdownMenuLabel>
+              <DropdownMenuSeparator />
+              <DropdownMenuItem onClick={onSignOut}>
+                <LogOut />
+                Sign out
+              </DropdownMenuItem>
+            </DropdownMenuGroup>
+          </DropdownMenuContent>
+        </DropdownMenu>
+      </SidebarMenuItem>
+    </SidebarMenu>
+  );
+}
